Pathfinding fixed
I have stopped using Node as a way to calculate the best way using my A*, this prevented the memory issues that were being had with the code. The pair used in the priority queue had the order swapped as this was causing the queue to compare the pointer of the waypoints, not the double that signified the cost of it.